requirements.txt
python-dotenv
langchain
PyPDF2
faiss-cpu
streamlit
huggingface_hub==0.25.1
langchain-community
requests==2.32.3
sentence-transformers==3.2.0
-e .
